[DEV] relative_symlink: compute relative to dst realpath
Use dst realpath when computing the relative path of the
link target.

This allows for the link target directory to be itself
a symlink.

For instance, this allows to copy additional files to
the 'images' dir (which is a symlink in the release dir
to the 'images' dir in out dir) using product_config.json
'additional_files' mechanism.