[DEV] md5sum.txt: support unusual filenames
The existing command for generating the
md5sum.txt file of release archives fails
when trying to compute the md5sum of files
which have an unusual name, such as files
with whitespaces or newlines in their
relative path.

Using find -print0 | xargs -0 combo allows
to handle such files gracefully.